feat(x2/x3): conditional attributes
Implements encoding of Conditional Attributes as TLV Objects. We've also
changed the API so that objects write their contents to a
DataOutputStream. We find this to be more clean and this is inspired
from Graphics APIs where shapes also can draw themselves onto surfaces.

Decodation is still broken and will be adressed in the next commit

Co-authored-by: Sven Uhlig <uhlig@sipgate.de>
Co-authored-by: Nick Garczorz <garczorz@sipgate.de>