fix(decoder): determine TLV length correctly
The MANDATORY_HEADER_LENGTH of 40 bytes is the minimum viable header
length, which does NOT include conditional attributes. So in order to
determine if we have conditional attributes, we need to take the total
header length and subtract 40 bytes from it.

The previous implementation used readableBytes() from the buffer, which
only works as long as there is exactly one packet inside the buffer and
breaks if there is more than one.

Co-authored-by: Nick Garczorz <garczorz@sipgate.de>
Co-authored-by: Sven Uhlig <uhlig@sipgate.de>