feat: provide complete clients / servers
While developing our internal LI solution, we noticed that there are a
lot of things that would also be beneficial in a public library. So we
have moved all code that can be public to this library. This mainly
includes the server handling of X1/X2/X3 through netty as well as and a
lot of X1 buissness logic of what happens when creating tasks /
destinations and so on.

As this increased the codebase by a lot, we have re-organized all
packages as well.

Co-Authored-By: Lennart Rosam <rosam@sipgate.de>