fix(x1server): improve keep-alive handlings
When we respond with an error, we want to close the connection. However
we must also send a "Connection: close" header in the response when we
intend to close it. Otherwise, clients might assume that the connection
can be re-used and will fail.

The joys of implementing your own HTTP server :)

Co-authored-by: Sven Uhlig <uhlig@sipgate.de>